+++ app/src/main/java/ru/evgeniy/dpitunnelcli/domain/usecases/IRenameProfileUseCase.kt
@@ -1,5 +1,5 @@
 package ru.evgeniy.dpitunnelcli.domain.usecases
 
 interface IRenameProfileUseCase {
-    suspend fun rename(id: Int, newTitle: String)
+    suspend fun rename(id: Long, newTitle: String)
 }
