+++ app/src/main/java/ru/evgeniy/dpitunnelcli/domain/usecases/ISettingsUseCase.kt
@@ -5,6 +5,6 @@ interface ISettingsUseCase {
     fun getCABundlePath(): String?
     fun getIP(): String?
     fun getPort(): Int?
-    fun getDefaultProfileId(): Int?
-    fun setDefaultProfileId(value: Int?)
+    fun getDefaultProfileId(): Long?
+    fun setDefaultProfileId(value: Long?)
 }
